# WebAPI/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WebAPI",
    "version": "v1"
  },
  "paths": {
    "/api/Account/authenticate": {
      "post": {
        "tags": [
          "Account"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AuthenticationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Account/register": {
      "post": {
        "tags": [
          "Account"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Recaudos": {
      "get": {
        "tags": [
          "Recaudos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "fechaRecaudo",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "estacion",
            "type": "string"
          },
          {
            "in": "query",
            "name": "sentido",
            "type": "string"
          },
          {
            "in": "query",
            "name": "categoria",
            "type": "string"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Recaudos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateRecaudoCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Recaudos/{id}": {
      "get": {
        "tags": [
          "Recaudos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Recaudos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateRecaudoCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Recaudos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "AuthenticationRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "RegisterRequest": {
      "type": "object",
      "properties": {
        "nombre": {
          "type": "string"
        },
        "apellido": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        }
      }
    },
    "CreateRecaudoCommand": {
      "type": "object",
      "properties": {
        "fechaRecaudo": {
          "format": "date-time",
          "type": "string"
        },
        "estacion": {
          "type": "string"
        },
        "sentido": {
          "type": "string"
        },
        "categoria": {
          "type": "string"
        },
        "hora": {
          "format": "int32",
          "type": "integer"
        },
        "valorTabulado": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "UpdateRecaudoCommand": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "fechaRecaudo": {
          "format": "date-time",
          "type": "string"
        },
        "estacion": {
          "type": "string"
        },
        "sentido": {
          "type": "string"
        },
        "categoria": {
          "type": "string"
        },
        "hora": {
          "format": "int32",
          "type": "integer"
        },
        "valorTabulado": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}